# PR

```
VLLM_LOGGING_LEVEL=DEBUG vllm serve deepseek-ai/DeepSeek-Coder-V2-Lite-Instruct
...
(EngineCore_DP0 pid=3662669) DEBUG 10-01 17:41:11 [v1/attention/.../mla/common.py:980] Using FlashInfer prefill for MLA
...
========================================================================
(vllm) lwilkinson@dgxB200-09:~/code/vllm$ python tests/evals/gsm8k/gsm8k_eval.py
Running GSM8K evaluation: 1319 questions, 5-shot
Evaluating: 100%|█████████████████████████████████████████████████████████████████████| 1319/1319 [00:28<00:00, 46.50it/s]

Results:
Accuracy: 0.782
Invalid responses: 0.000
Total latency: 28.381 s
Questions per second: 46.475
```

# Main

```
(vllm) lwilkinson@dgxB200-09:~/code/vllm$ python tests/evals/gsm8k/gsm8k_eval.py
Running GSM8K evaluation: 1319 questions, 5-shot
Evaluating: 100%|█████████████████████████████████████████████████████████████████████| 1319/1319 [00:26<00:00, 49.12it/s]

Results:
Accuracy: 0.208
Invalid responses: 0.006
Total latency: 26.864 s
Questions per second: 49.100
```