@benchislett 2025-10-02
I ran this to validate on 8xB200:
```
VLLM_ATTENTION_BACKEND=CUTLASS_MLA vllm serve deepseek-ai/DeepSeek-R1-0528 -tp 8 --max-model-len 8192 --no-enable-prefix-caching --port 8049 --max-num-batched-tokens 256
```

where it completely failed before, it passes GSM8k now.

I also ran it with MTP on my development branch (#25984), which relies more heavily on this prefill functionality, and it passed as well:

```
VLLM_ATTENTION_BACKEND=CUTLASS_MLA vllm serve deepseek-ai/DeepSeek-R1-0528 -tp 8 --max-model-len 8192 --no-enable-prefix-caching --port 8049 --max-num-batched-tokens 256 --speculative-config '{"method": "mtp", "num_speculative_tokens": 3}'

|Tasks|Version|     Filter     |n-shot|  Metric   |   |Value |   |Stderr|
|-----|------:|----------------|-----:|-----------|---|-----:|---|-----:|
|gsm8k|      3|flexible-extract|     5|exact_match|↑  |0.9538|±  |0.0058|
|     |       |strict-match    |     5|exact_match|↑  |0.9500|±  |0.0060|
```